+++ NEWS.md
@@ -1,5 +1,13 @@
 # guideR (development version)
 
+**New feature**
+
+* new function `plot_multiple_answers_dodge()` (#31)
+
+**Fix**
+
+* fix in `plot_multiple_answers()` with `NA` values when `drop_na = FALSE` (#31)
+
 # guideR 0.5.0
 
 **New features**